refactor: modularize layout and admin components
